test(generators): adopt the review probe as HintNameCollisionTests; drop the scratch consumer
The adversarial review left an empirical probe (raw Assert, scratch consumer
project) in the worktree, which the previous commit swept up. The probe's
coverage is kept — whole-driver collision runs and the first-line
column-addition branch of the block-to-file composition — rewritten to the
repo's Shouldly/naming/testing conventions; the scratch consumer project is
removed.
