--- scratch-review-consumer/Class1.cs
@@ -1,5 +1,0 @@
-namespace ScratchReviewConsumer;
-
-public sealed class Class1
-{
-}
