--- scratch-review-consumer/Hello.viu
@@ -1,3 +1,0 @@
-@template {
-  <div>{{ message }}</div>
-}
