--- scratch-review-consumer/scratch-review-consumer.csproj
@@ -1,12 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <TargetFramework>$(TargetFrameworkForLibraries)</TargetFramework>
-    <IsPackable>false</IsPackable>
-  </PropertyGroup>
-  <ItemGroup>
-    <VuecsAnalyzerReference Include="Assimalign.Vue.Syntax.Generators" />
-    <AdditionalFiles Include="Hello.viu" />
-    <CompilerVisibleProperty Include="RootNamespace" />
-    <CompilerVisibleProperty Include="ProjectDir" />
-  </ItemGroup>
-</Project>
